+++ firmware/init/sensor/init_sensors.cpp
@@ -129,7 +129,19 @@ void reconfigureSensors() {
 
 // Mocking/testing helpers
 static void initSensorCli() {
-	addConsoleActionIF("set_sensor_mock", Sensor::setMockValue);
+	addConsoleActionSS("set_sensor_mock", [](const char* typeName, const char* valueStr) {
+		SensorType type = findSensorTypeByName(typeName);
+
+		if (type == SensorType::Invalid) {
+			efiPrintf("Invalid sensor type specified: %s", typeName);
+			return;
+		}
+
+		float value = atoff(valueStr);
+
+		Sensor::setMockValue(type, value);
+	});
+
 	addConsoleAction("reset_sensor_mocks", Sensor::resetAllMocks);
 	addConsoleAction("show_sensors", Sensor::showAllSensorInfo);
 	addConsoleActionI("show_sensor",